Shift LiquidatedVaultDerivedValues interface
